refactor(settings): split panel sections
